Fix Prisma scripts in package.json, make logic flow clearer
