+++ nodemon.json
@@ -1,6 +1,6 @@
 {
     "ignore": ["**/*.test.ts", "**/*.spec.ts", ".git", "node_modules"],
     "watch": ["dist"],
-    "exec": "node --experimental-json-modules dist/src/index.js",
+    "exec": "npm run env:set-database-url -- node --experimental-json-modules dist/src/scripts/start.js",
     "ext": "js"
 }
