+++ package.json
@@ -24,10 +24,12 @@
     "build": "tsc",
     "build:watch": "tsc --watch",
     "prepack": "npm run clean && npm run build",
-    "start": "node --experimental-json-modules dist/src/scripts/prestart.js",
+    "start": "npm run env:set-database-url -- node --experimental-json-modules dist/src/scripts/migrate-and-start.js",
     "dev": "concurrently nodemon 'tsc --watch'",
-    "migrations:generate": "prisma migrate dev",
-    "migrations:run": "prisma migrate deploy",
+    "migrations:generate": "npm run prisma migrate dev",
+    "migrations:run": "npm run prisma migrate deploy",
+    "prisma": "npm run env:set-database-url prisma",
+    "env:set-database-url": "node dist/src/scripts/run-with-database-url.js",
     "release": "release-it"
   },
   "devDependencies": {
