+++ src/index.ts
@@ -1,11 +1,24 @@
-import {DATA_DIR} from './services/config.js';
-import {startBot} from './scripts/start.js';
-import createDatabaseUrl from './utils/create-database-url.js';
-import logBanner from './utils/log-banner.js';
+import makeDir from 'make-dir';
+import path from 'path';
+import container from './inversify.config.js';
+import {TYPES} from './types.js';
+import Bot from './bot.js';
+import Config from './services/config.js';
+import FileCacheProvider from './services/file-cache.js';
 
-process.env.DATABASE_URL = createDatabaseUrl(DATA_DIR);
+const bot = container.get<Bot>(TYPES.Bot);
 
-(async () => {
-  logBanner();
-  await startBot();
-})();
+const startBot = async () => {
+  // Create data directories if necessary
+  const config = container.get<Config>(TYPES.Config);
+
+  await makeDir(config.DATA_DIR);
+  await makeDir(config.CACHE_DIR);
+  await makeDir(path.join(config.CACHE_DIR, 'tmp'));
+
+  await container.get<FileCacheProvider>(TYPES.FileCache).cleanup();
+
+  await bot.listen();
+};
+
+export {startBot};
