+++ src/scripts/start.ts
@@ -1,24 +1,9 @@
-import makeDir from 'make-dir';
-import path from 'path';
-import container from '../inversify.config.js';
-import {TYPES} from '../types.js';
-import Bot from '../bot.js';
-import Config from '../services/config.js';
-import FileCacheProvider from '../services/file-cache.js';
+// This script is mainly used during development.
+// Starts Muse without applying database migrations.
+import {startBot} from '../index.js';
+import logBanner from '../utils/log-banner.js';
 
-const bot = container.get<Bot>(TYPES.Bot);
-
-const startBot = async () => {
-  // Create data directories if necessary
-  const config = container.get<Config>(TYPES.Config);
-
-  await makeDir(config.DATA_DIR);
-  await makeDir(config.CACHE_DIR);
-  await makeDir(path.join(config.CACHE_DIR, 'tmp'));
-
-  await container.get<FileCacheProvider>(TYPES.FileCache).cleanup();
-
-  await bot.listen();
-};
-
-export {startBot};
+(async () => {
+  logBanner();
+  await startBot();
+})();
